MiniTensor: rename headers to match their contents
Three headers carried names inherited from older codes that no longer
describe what they contain:

- MiniTensor_Definitions.h -> MiniTensor_Traits.h: the scalar/AD
  type-traits layer (Sacado promotion machinery, index types).
- MiniTensor_Utilities.h -> MiniTensor_Scalar.h: scalar-level helpers
  (machine_epsilon, not_a_number, random, sgn, ...).
- MiniTensor_FunctionSet.h -> MiniTensor_TestFunctions.h: benchmark
  objective functions for exercising the solvers.

The old names remain as deprecated forwarding headers that simply
include the new ones, so existing users compile unchanged. Internal
includes are updated to the new names. Include guards follow the new
names. Two latent gaps are fixed along the way: the test-functions
header gains the include guard it previously lacked, and the scalar
header now includes MiniTensor_Traits.h for the Index type it uses
instead of relying on include order in the TensorBase chain.
